requirements.txt
psycopg2
pydantic
pytest-postgresql